fix also for other PipeItems
